// frontend/src/api/models/handlerAutoPrintRuleHTTPResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Auto print rule response
 */
export interface HandlerAutoPrintRuleHTTPResponse {
  auto_print?: boolean
  copies?: number
  created_at?: string
  document_type?: string
  enabled?: boolean
  id?: string
  printer_name?: string
  template_id?: string
  tenant_id?: string
  trigger_event?: string
  updated_at?: string
}