// frontend/src/api/models/handlerCreateAutoPrintRuleHTTPRequest.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Request body for creating an auto print rule
 */
export interface HandlerCreateAutoPrintRuleHTTPRequest {
  auto_print?: boolean
  /**
   * @minimum 1
   * @maximum 100
   */
  copies?: number
  document_type: string
  printer_name?: string
  template_id?: string
  trigger_event: string
}